+++ udata_recommendations/__init__.py
@@ -1,9 +1,6 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
 '''
 uData recommendations
 '''
-from __future__ import unicode_literals
 
-__version__ = '1.0.2.dev'
+__version__ = '2.0.0.dev'
 __description__ = 'uData content recommendations bridge'