+++ udata_recommendations/views.py
@@ -1,6 +1,3 @@
-# -*- coding: utf-8 -*-
-from __future__ import unicode_literals
-
 from flask import Blueprint
 
 from udata import theme
@@ -11,8 +8,14 @@ from udata.frontend import template_hook
 blueprint = Blueprint('recommendations', __name__, template_folder='templates')
 
 
-@template_hook
-def dataset_recommendations(reco_ids):
+def has_recommendations(ctx):
+    dataset = ctx['dataset']
+    return 'recommendations' in dataset.extras and len(dataset.extras['recommendations'])
+
+
+@template_hook('dataset.display.after-description', when=has_recommendations)
+def dataset_recommendations(ctx):
+    reco_ids = ctx['dataset'].extras['recommendations']
     recommendations = Dataset.objects.filter(id__in=reco_ids)
     return theme.render('dataset-recommendations.html',
                         recommendations=recommendations)